Fixed labels for Mil2525CMessageLayer; updated to ArcGIS 10.2.3
Fixes #18

+++ source/VehicleCommander/src/com/esri/vehiclecommander/view/IdentifyResultsJPanel.java
@@ -390,7 +390,7 @@ public class IdentifyResultsJPanel extends RoundedJPanel implements LocationList
                 public void actionPerformed(ActionEvent e) {
                     BrowserLauncher.openURL(featureLayer.getAttachmentURL(
                             Integer.parseInt((String) result.getAttributes().get(featureLayer.getObjectIdField())),
-                            attInfo.getId()));
+                            (int) attInfo.getId()));
                 }
             });
             jPanel_results.add(button, c);
